refactor: move the logic gadgets into composer::logic
Pure code motion: `append_logic_component`, its two binding helpers and
the `append_logic_and`/`append_logic_xor` entry points move out of
`composer.rs` into a `logic` submodule, together with their soundness
tests.
